feat(domain): 在 LoginBo 中添加用户类型和租户 ID 字段
- 在 LoginBo 中添加 userType 和 tenantId 字段，并使用 @Schema 注解描述字段信息
- 优化 NotificationServiceImpl 中的 getUnreadNotificationCount 方法，确保返回类型正确
- 调整 pom.xml 中的依赖项格式

+++ src/main/java/com/dsy/hangaituangou/service/impl/NotificationServiceImpl.java
@@ -82,7 +82,7 @@ public class NotificationServiceImpl extends ServiceImpl<NotificationMapper, Not
      */
     @Override
     public int getUnreadNotificationCount(Long receiverId) {
-        return count(new LambdaQueryWrapper<Notification>()
+        return (int) count(new LambdaQueryWrapper<Notification>()
                 .eq(Notification::getReceiverId, receiverId)
                 .eq(Notification::getStatus, 0));
     }
